requirements.txt
requests
pillow
matplotlib
pandas